move test for sanitization helper to separate test file
